code/requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning==1.8.1
torchmetrics==0.10.0

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)

opencv-python==4.5.5.62
opencv-contrib-python==4.5.5.62
pandas
matplotlib
seaborn
scikit-image
scikit-learn
timm


# --------- deployment --------- #
streamlit==1.4.0
boto3==1.20.37
Pillow==9.0.0
pandas==1.3.5
tqdm==4.62.3
